requirements.txt
requests~=2.28.1
selenium~=4.3.0
awscli
# pip install -r requirements.txt
Flask~=2.1.3
Pillow~=9.2.0
inputimeout~=1.0.4
names~=0.3.0